project/requirements/base.txt
django-crispy-forms==1.9.2
django-environ==0.4.5
django==3.2.10
easy-thumbnails==2.7.1
pillow==8.0.1
pytz==2020.4
django_tables2==2.3.3
django-role-permissions==2.2.1
django-extensions==2.1.5
djangorestframework==3.12.4
gunicorn==19.9.0
django-guardian
uwsgi
django-tinymce
djangorestframework-simplejwt
djoser
pytest
pytest-django
pytest-pythonpath
pytest-cov
mixer